Refactor controllers and automapper to implement UnitOfWork pattern

+++ ClinicManagement/Models/AppointmentCreateViewModel.cs
@@ -11,7 +11,7 @@ namespace ClinicManagement.Models
     {
         public string DoctorId { get; set; }
         public DoctorViewModel Doctor { get; set; }
-        public int? SpecialtyId { get; set; }
+        public long? SpecialtyId { get; set; }
         public SpecialtyViewModel Specialty { get; set; }
     }
 }
